pyproject.toml
[project]
name = "website-with-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "google-generativeai>=0.8.5",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
]
